Feature/letter

+++ src/main/java/com/example/tree/tree/TreeService.java
@@ -2,7 +2,10 @@ package com.example.tree.tree;
 
 import com.example.tree.User.User;
 import com.example.tree.User.UserRepository;
+import com.example.tree.letter.domain.LetterDao;
 import com.example.tree.letter.dto.LetterSimpleResponse;
+import com.example.tree.letter.exception.AccessDeniedException;
+import com.example.tree.tree.exception.TreeNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -14,31 +17,47 @@ public class TreeService {
     private final TreeRepository treeRepository;
     private final UserRepository userRepository;
     private final TreeDao treeDao;
+    private final LetterDao letterDao;
 
-    public TreeService(TreeDao treeDao, TreeRepository treeRepository, UserRepository userRepository, UserRepository userRepository1) {
+    public TreeService(TreeDao treeDao, LetterDao letterDao, TreeRepository treeRepository, UserRepository userRepository, UserRepository userRepository1) {
         this.treeDao = treeDao;
         this.treeRepository = treeRepository;
         this.userRepository = userRepository1;
-    }
-    public void create(createTreeRequest request) {
-        User user = userRepository.findById(request.userId()).orElseThrow(() -> new NoSuchElementException("userId가 없습니다."));
-        treeRepository.save(new Tree(request.title(),user));
+        this.letterDao = letterDao;
     }
 
-    public TreeResponse read(Long treeId) {
+    public void create(createTreeRequest request, Long userId) {
+        User user = userRepository.findById(userId).orElseThrow(() -> new NoSuchElementException("userId가 없습니다."));
+        treeRepository.save(new Tree(request.title(), user));
+    }
+
+    public TreeResponse read(Long treeId, String order) {
         Tree tree = treeDao.findByTreeIdAndIsOpen(treeId);
-        List<LetterSimpleResponse> letterSRPList = tree.getLetterList().stream()
+        List<LetterSimpleResponse> letterSRPList = letterDao.findAllByTreeIdOrderBy(treeId, order).stream()
                 .map(LetterSimpleResponse::toDto)
                 .toList();
-        return new TreeResponse(tree.getId(),tree.getTitle(),tree.getUrlPath(),tree.isOpen(),tree.getUser().getId(),letterSRPList);
+        return new TreeResponse(tree.getId(), tree.getTitle(), tree.getUrlPath(), tree.isOpen(), tree.getUser().getId(), letterSRPList);
     }
+
     @Transactional
-    public void delete(Long treeId) {
-        treeRepository.deleteById(treeId);
+    public void delete(Long treeId, Long userId) {
+        Tree tree = treeRepository.findById(treeId).orElseThrow(()->new TreeNotFoundException("존재하지 않는 트리입니다."));
+        if (tree.getUser().getId().equals(userId)){
+            treeRepository.deleteById(treeId);
+        }
+        else {
+            throw new AccessDeniedException("삭제 권한이 없습니다.");
+        }
     }
+
     @Transactional
-    public void close(Long treeId) {
+    public void close(Long treeId, Long userId) {
         Tree tree = treeRepository.findById(treeId).orElseThrow(() -> new NoSuchElementException("id가 없습니다."));
-        tree.isClose();
+        if (tree.getUser().getId().equals(userId)){
+            tree.isClose();
+        }
+        else {
+            throw new AccessDeniedException("비공개 권한이 없습니다.");
+        }
     }
 }
